sync(wait): towards a general execution space based sync wait (for graph too)
