+++ CMakeLists.txt
@@ -112,6 +112,7 @@ target_sources(
             kokkos-execution/impl/HPX/event.hpp
             kokkos-execution/impl/sender_concepts.hpp
             kokkos-execution/impl/SYCL/event.hpp
+            kokkos-execution/impl/state.hpp
             kokkos-execution/impl/sync_wait.hpp
             kokkos-execution/utils/ignore_warnings.hpp
 )
